Added new 'AssetCompiler' interface. * Assets loaders will now expect 'compiled' or 'baked' assets. * A loaders task is to fixup any pointers and use the input resource as is as much as possible. * A compilers task is to transform any format to the engine asset format.

+++ source/code/core/core/public/core/cexpr/stringid.hxx
@@ -45,7 +45,7 @@ namespace core
             int32_t const copy_count = std::min(origin_size, cstr_size);
             int32_t const copy_offset = std::max(0, cstr_size - copy_count);
 
-            int32_t i = 0;
+            int64_t i = 0;
             for (auto& v : result.hash_origin)
             {
                 if (i < copy_count)
